requirements.txt
Django==4.2.17
psycopg2-binary==2.9.10
djangorestframework==3.15.*
djangorestframework-simplejwt==5.3.*
pillow==11.0.*
django-extensions==3.2.*
